+++ music/Freak.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+
+song_of_the_day_name="Freak"
+song_of_the_day_album="Honeymoon"
+song_of_the_day_artist="Lana Del Rey"
+song_of_the_day_spotify="https://open.spotify.com/track/3sYDVtqO35oRSOIMx7dOqR?si=c82200d38e864d8a"
+song_of_the_day_youtube="https://youtu.be/jq30l5-vBbo?si=4FxjqOZYKTOBtaQp"
+song_of_the_day_quote="Oh, you're cold as ice baby"